자바게시판 수정

+++ src/service/BoardService.java
@@ -30,19 +30,20 @@ public class BoardService {
             System.out.println("글쓰기 실패");
         }
     }
+
     //글조회
     public void findById() {
         System.out.print("조회할 id: ");
-        long id = scanner.nextLong();
-        BoardDTO boardDTO= boardRepository.findById(id);
+        Long id = scanner.nextLong();
+        BoardDTO boardDTO = boardRepository.findById(id);
         //System.out.println(boardDTO.toString());
         if (boardDTO != null) {
-            boardDTO.setBoardHits(boardDTO.getBoardHits()+1);
+            boardDTO.setBoardHits(boardDTO.getBoardHits() + 1);
             System.out.println("글번호 : " + boardDTO.getId());
             System.out.println("제목 : " + boardDTO.getBoardTitle());
             System.out.println("작성자 : " + boardDTO.getBoardWriter());
-            System.out.println("조회수 : " + boardDTO.getBoardHits() +" 회");
-            System.out.println("내용 : " +  boardDTO.getBoardContents());
+            System.out.println("조회수 : " + boardDTO.getBoardHits() + " 회");
+            System.out.println("내용 : " + boardDTO.getBoardContents());
         } else {
             System.out.println("찾는 글이 없습니다");
         }
@@ -50,13 +51,13 @@ public class BoardService {
 
     //글목록
     public void findAll() {
-        List<BoardDTO>  boardDTOList = boardRepository.findAll();
+        List<BoardDTO> boardDTOList = boardRepository.findAll();
         for (BoardDTO boardDTO : boardDTOList) {
             //System.out.println(boardDTO.toString());
             System.out.println("글번호 : " + boardDTO.getId());
             System.out.println("제목 : " + boardDTO.getBoardTitle());
             System.out.println("작성자 : " + boardDTO.getBoardWriter());
-            System.out.println("조회수 : " +  boardDTO.getBoardHits());
+            System.out.println("조회수 : " + boardDTO.getBoardHits());
             System.out.println("-------------------------------");
         }
     }
@@ -69,19 +70,23 @@ public class BoardService {
         String password = scanner.next();
 
         BoardDTO boardDTO = boardRepository.findById(id);
-        if(password.equals(boardDTO.getBoardPass())) {
-            System.out.print("수정할 제목 입력:   ");
-            String title = scanner.next();
-            System.out.print("수정할 내용 입력:   ");
-            String contents = scanner.next();
-            boolean result = boardRepository.update(id, title, contents);
-            if (result) {
-                System.out.println("수정 완료");
+        if (boardDTO != null) {
+            if (password.equals(boardDTO.getBoardPass())) {
+                System.out.print("수정할 제목 입력:   ");
+                String title = scanner.next();
+                System.out.print("수정할 내용 입력:   ");
+                String contents = scanner.next();
+                boolean result = boardRepository.update(id, title, contents);
+                if (result) {
+                    System.out.println("수정 완료");
+                } else {
+                    System.out.println("수정 실패");
+                }
             } else {
-                System.out.println("수정 실패");
+                System.out.println("비밀번호 오류");
             }
         } else {
-            System.out.println("비밀번호 오류");
+            System.out.println("해당 내용이 없습니다");
         }
     }
 
@@ -90,14 +95,14 @@ public class BoardService {
         System.out.print("검색할 제목 입력:  ");
         String title = scanner.next();
 
-        List<BoardDTO>  boardDTOList1 = boardRepository.search(title);
+        List<BoardDTO> boardDTOList1 = boardRepository.search(title);
         if (boardDTOList1 != null) {
-        for (BoardDTO boardDTO : boardDTOList1) {
-            System.out.println("글번호 : " + boardDTO.getId() + "   작성자 : " + boardDTO.getBoardWriter());
-            System.out.println(boardDTO.getBoardContents());
-            System.out.println("==============================");
-        }
-        }else{
+            for (BoardDTO boardDTO : boardDTOList1) {
+                System.out.println("글번호 : " + boardDTO.getId() + "   작성자 : " + boardDTO.getBoardWriter());
+                System.out.println(boardDTO.getBoardContents());
+                System.out.println("==============================");
+            }
+        } else {
             System.out.println("해당 내용이 없습니다");
         }
     }
@@ -111,12 +116,16 @@ public class BoardService {
         String password = scanner.next();
 
         BoardDTO boardDTO = boardRepository.findById(id);
-        if(password.equals(boardDTO.getBoardPass())) {
-            boolean del = boardRepository.delete(id);
-            if(del) System.out.println("삭제 완료");
-            else System.out.println("해당 정보가 없습니다");
+        if (boardDTO != null) {
+            if (password.equals(boardDTO.getBoardPass())) {
+                boardRepository.delete(id);
+                System.out.println("삭제 완료");
+            } else {
+                System.out.println("비밀번호 오류");
+            }
         } else {
-            System.out.println("비밀번호 오류");
+            System.out.println("해당 정보가 없습니다");
         }
     }
 }
+
